move TestLuaScript to custom_network_provider_test
