+++ lua_configuration/convert_test_case_to_lua_object.go
@@ -29,13 +29,13 @@ type TestCase struct {
 // run `go run lua.go`, then this program will get all testdata and convert them into lua objects
 // copy the generated objects to lua scripts and then you can start debugging your lua scripts
 func main() {
-	err := PathWalk()
+	err := convertTestCaseToLuaObject()
 	if err != nil {
 		fmt.Println(err)
 	}
 }
 
-func PathWalk() error {
+func convertTestCaseToLuaObject() error {
 	err := filepath.Walk("./", func(path string, f os.FileInfo, err error) error {
 		if !strings.Contains(path, "trafficRouting.lua") {
 			return nil
@@ -51,7 +51,7 @@ func PathWalk() error {
 		err = filepath.Walk(filepath.Join(dir, "testdata"), func(path string, info os.FileInfo, err error) error {
 			if !info.IsDir() && filepath.Ext(path) == ".yaml" || filepath.Ext(path) == ".yml" {
 				fmt.Printf("--- walking path: %s ---\n", path)
-				err = ObjectToTable(path)
+				err = objectToTable(path)
 				if err != nil {
 					return fmt.Errorf("failed to convert object to table: %s", err)
 				}
@@ -70,7 +70,7 @@ func PathWalk() error {
 }
 
 // convert a testcase object to lua table for debug
-func ObjectToTable(path string) error {
+func objectToTable(path string) error {
 	dir, file := filepath.Split(path)
 	testCase, err := getLuaTestCase(path)
 	if err != nil {
@@ -138,7 +138,8 @@ func ObjectToTable(path string) error {
 		return fmt.Errorf("failed to open file: %s", err)
 	}
 	defer fileStream.Close()
-	_, err = io.WriteString(fileStream, objStr)
+	header := "-- THIS IS GENERATED BY LUA.GO FOR DEBUGGING --\n"
+	_, err = io.WriteString(fileStream, header+objStr)
 	if err != nil {
 		return fmt.Errorf("failed to WriteString %s", err)
 	}
