+++ pkg/util/luamanager/lua_test.go
@@ -19,7 +19,6 @@ package luamanager
 import (
 	"encoding/json"
 	"fmt"
-	"os"
 	"testing"
 
 	rolloutv1alpha1 "github.com/openkruise/rollouts/api/v1alpha1"
@@ -27,26 +26,17 @@ import (
 	lua "github.com/yuin/gopher-lua"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
-	utilpointer "k8s.io/utils/pointer"
 	luajson "layeh.com/gopher-json"
 	gatewayv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
-
-	"path/filepath"
-	"strings"
-
-	"sigs.k8s.io/yaml"
 )
 
 type LuaData struct {
-	Data             Data
-	CanaryWeight     int32
-	StableWeight     int32
-	Matches          []rolloutv1alpha1.HttpRouteMatch
-	CanaryService    string
-	StableService    string
-	RevisionLabelKey string
-	StableRevision   string
-	CanaryRevision   string
+	Data          Data
+	CanaryWeight  int32
+	StableWeight  int32
+	Matches       []rolloutv1alpha1.HttpRouteMatch
+	CanaryService string
+	StableService string
 }
 type Data struct {
 	Spec        interface{}       `json:"spec,omitempty"`
@@ -161,177 +151,4 @@ func TestRunLuaScript(t *testing.T) {
 			}
 		})
 	}
-}
-
-type LuaTestCase struct {
-	Matches       []rolloutv1alpha1.HttpRouteMatch `yaml:"matches"`
-	StableService string                           `yaml:"stableService"`
-	CanaryService string                           `yaml:"canaryService"`
-	StableWeight  int                              `yaml:"stableWeight"`
-	CanaryWeight  int                              `yaml:"canaryWeight"`
-	Spec          interface{}                      `yaml:"spec"`
-	NSpec         interface{}                      `yaml:"nSpec"`
-}
-
-type TestCase struct {
-	Rollout        *rolloutv1alpha1.Rollout        `json:"rollout,omitempty"`
-	TrafficRouting *rolloutv1alpha1.TrafficRouting `json:"trafficRouting,omitempty"`
-	Original       *unstructured.Unstructured      `json:"original,omitempty"`
-	Expected       []*unstructured.Unstructured    `json:"expected,omitempty"`
-}
-
-// test if the lua script run as expected
-func TestLuaScript(t *testing.T) {
-	err := filepath.Walk("../../../lua_configuration", func(path string, f os.FileInfo, err error) error {
-		if !strings.Contains(path, "trafficRouting.lua") {
-			return nil
-		}
-		if err != nil {
-			return err
-		}
-		script, err := readScript(t, path)
-		if err != nil {
-			return err
-		}
-		dir := filepath.Dir(path)
-		err = filepath.Walk(filepath.Join(dir, "testdata"), func(path string, info os.FileInfo, err error) error {
-			if err != nil {
-				return err
-			}
-
-			if !info.IsDir() && filepath.Ext(path) == ".yaml" || filepath.Ext(path) == ".yml" {
-				testCase := getLuaTestCase(t, path)
-				rollout := testCase.Rollout
-				trafficRouting := testCase.TrafficRouting
-				if rollout != nil {
-					steps := rollout.Spec.Strategy.Canary.Steps
-					for i, step := range steps {
-						weight := step.TrafficRoutingStrategy.Weight
-						if weight == nil {
-							weight = utilpointer.Int32(-1)
-						}
-						var canaryService string
-						stableService := rollout.Spec.Strategy.Canary.TrafficRoutings[0].Service
-						canaryService = fmt.Sprintf("%s-canary", stableService)
-						data := &LuaData{
-							Data: Data{
-								Labels:      testCase.Original.GetLabels(),
-								Annotations: testCase.Original.GetAnnotations(),
-								Spec:        testCase.Original.Object["spec"],
-							},
-							Matches:          step.TrafficRoutingStrategy.Matches,
-							CanaryWeight:     *weight,
-							StableWeight:     100 - *weight,
-							CanaryService:    canaryService,
-							StableService:    stableService,
-							RevisionLabelKey: "pod-template-hash",
-							StableRevision:   "pod-template-hash-stable",
-							CanaryRevision:   "pod-template-hash-canary",
-						}
-						nSpec, err := executeLua(data, script)
-						if err != nil {
-							return err
-						}
-						eSpec := Data{
-							Spec:        testCase.Expected[i].Object["spec"],
-							Annotations: testCase.Expected[i].GetAnnotations(),
-							Labels:      testCase.Expected[i].GetLabels(),
-						}
-						if util.DumpJSON(eSpec) != util.DumpJSON(nSpec) {
-							return fmt.Errorf("expect %s, but get %s", util.DumpJSON(eSpec), util.DumpJSON(nSpec))
-						}
-					}
-				} else if trafficRouting != nil {
-					weight := trafficRouting.Spec.Strategy.Weight
-					if weight == nil {
-						weight = utilpointer.Int32(-1)
-					}
-					var canaryService string
-					stableService := trafficRouting.Spec.ObjectRef[0].Service
-					canaryService = stableService
-					data := &LuaData{
-						Data: Data{
-							Labels:      testCase.Original.GetLabels(),
-							Annotations: testCase.Original.GetAnnotations(),
-							Spec:        testCase.Original.Object["spec"],
-						},
-						Matches:          trafficRouting.Spec.Strategy.Matches,
-						CanaryWeight:     *weight,
-						StableWeight:     100 - *weight,
-						CanaryService:    canaryService,
-						StableService:    stableService,
-						RevisionLabelKey: "pod-template-hash",
-						StableRevision:   "pod-template-hash-stable",
-						CanaryRevision:   "pod-template-hash-canary",
-					}
-					nSpec, err := executeLua(data, script)
-					if err != nil {
-						return err
-					}
-					eSpec := Data{
-						Spec:        testCase.Expected[0].Object["spec"],
-						Annotations: testCase.Expected[0].GetAnnotations(),
-						Labels:      testCase.Expected[0].GetLabels(),
-					}
-					if util.DumpJSON(eSpec) != util.DumpJSON(nSpec) {
-						return fmt.Errorf("expect %s, but get %s", util.DumpJSON(eSpec), util.DumpJSON(nSpec))
-					}
-				} else {
-					return fmt.Errorf("neither rollout nor trafficRouting defined in test case: %s", path)
-				}
-			}
-			return nil
-		})
-		return err
-	})
-	if err != nil {
-		t.Fatalf("failed to test lua scripts: %s", err.Error())
-	}
-}
-
-func readScript(t *testing.T, path string) (string, error) {
-	data, err := os.ReadFile(filepath.Clean(path))
-	if err != nil {
-		return "", err
-	}
-	return string(data), err
-}
-
-func getLuaTestCase(t *testing.T, path string) *TestCase {
-	yamlFile, err := os.ReadFile(path)
-	if err != nil {
-		t.Fatalf("failed to read file %s", path)
-	}
-	luaTestCase := &TestCase{}
-	err = yaml.Unmarshal(yamlFile, luaTestCase)
-	if err != nil {
-		t.Fatalf("test case %s format error", path)
-	}
-	return luaTestCase
-}
-
-func executeLua(data *LuaData, script string) (Data, error) {
-	luaManager := &LuaManager{}
-	unObj, err := runtime.DefaultUnstructuredConverter.ToUnstructured(data)
-	if err != nil {
-		return Data{}, err
-	}
-	u := &unstructured.Unstructured{Object: unObj}
-	l, err := luaManager.RunLuaScript(u, script)
-	if err != nil {
-		return Data{}, err
-	}
-	returnValue := l.Get(-1)
-	var nSpec Data
-	if returnValue.Type() == lua.LTTable {
-		jsonBytes, err := luajson.Encode(returnValue)
-		if err != nil {
-			return Data{}, err
-		}
-		err = json.Unmarshal(jsonBytes, &nSpec)
-		if err != nil {
-			return Data{}, err
-		}
-	}
-	return nSpec, nil
 }
